Refactored code & folders
