--- src/containers/home/home.container.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-import { SafeArea } from '@/components/ui';
-import AddTodoForm from './add-todo-form.component';
-import HeaderSection from './header-section.component';
-import TodoListSection from './todo-list-section.component';
-
-const HomeContainer = () => {
-  return (
-    <SafeArea>
-      <HeaderSection />
-      <TodoListSection />
-      <AddTodoForm />
-    </SafeArea>
-  );
-};
-
-export default HomeContainer;
