--- src/containers/index.ts
@@ -1,2 +1,0 @@
-export { default as HomeContainer } from './home/home.container';
-export { default as SplashContainer } from './splash/splash.container';
