+++ src/features/screens.ts
@@ -1,0 +1,2 @@
+export { default as TodoListScreen } from './todo/screens/todo-list.screen';
+export { default as SplashScreen } from './splash/splash.screen';
