+++ src/features/todo/services/types.ts
@@ -1,5 +1,6 @@
 export interface Todo {
-  id: number;
+  id: string;
   title: string;
   completed: boolean;
+  timestamp: string;
 }
