+++ src/store/theme/index.ts
@@ -1,53 +1,3 @@
-import { createSlice } from '@reduxjs/toolkit';
-import themes from '@/config/theme/themes';
-import { RootState } from '..';
-
-const slice = createSlice({
-  name: 'theme',
-  initialState: { theme: 'default', darkMode: null } as ThemeState,
-  reducers: {
-    changeTheme: (state, { payload: { theme, darkMode } }: ThemePayload) => {
-      if (typeof theme !== 'undefined') {
-        state.theme = theme;
-      }
-      if (typeof darkMode !== 'undefined') {
-        state.darkMode = darkMode;
-      }
-    },
-    setDefaultTheme: (
-      state,
-      { payload: { theme, darkMode } }: ThemePayload,
-    ) => {
-      if (!state.theme) {
-        if (typeof theme !== 'undefined') {
-          state.theme = theme;
-        }
-        if (typeof darkMode !== 'undefined') {
-          state.darkMode = darkMode;
-        }
-      }
-    },
-  },
-});
-
-export const { changeTheme, setDefaultTheme } = slice.actions;
-
-export const selectThemes = (state: RootState) => state.theme;
-
-export default slice.reducer;
-
-type DarkProps<T> = {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  [K in keyof T]: K extends `${infer Prefix}_dark` ? K : never;
-}[keyof T];
-
-type PropsWithoutDark<T> = Omit<T, DarkProps<T>>;
-
-export type ThemeState = {
-  theme: 'default' | keyof PropsWithoutDark<typeof themes>;
-  darkMode: boolean | null;
-};
-
-type ThemePayload = {
-  payload: Partial<ThemeState>;
-};
+export { default as themeReducer } from './theme.slice';
+export * from './theme.slice';
+export * from './types';
