--- src/store/todo/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "todoSlice.ts"
-}
